+++ core/feed/annotator/loan_and_hold.py
@@ -2,7 +2,7 @@ import copy
 from datetime import datetime
 from typing import Any, Dict, List, Optional
 
-from core.feed_protocol.types import FeedData, Link, WorkEntry
+from core.feed.types import FeedData, Link, WorkEntry
 from core.model.configuration import ExternalIntegration
 from core.model.constants import EditionConstants, LinkRelations
 from core.model.patron import Hold, Patron
