+++ pyproject.toml
@@ -78,7 +78,7 @@ module = [
     "api.circulation",
     "api.discovery.*",
     "api.integration.*",
-    "core.feed_protocol.*",
+    "core.feed.*",
     "core.integration.*",
     "core.model.announcements",
     "core.model.hassessioncache",
