+++ tests/core/test_app_server.py
@@ -22,7 +22,7 @@ from core.app_server import (
 )
 from core.config import Configuration
 from core.entrypoint import AudiobooksEntryPoint, EbooksEntryPoint
-from core.feed_protocol.annotator.base import Annotator
+from core.feed.annotator.base import Annotator
 from core.lane import Facets, Pagination, SearchFacets, WorkList
 from core.log import LogConfiguration
 from core.model import ConfigurationSetting, Identifier
